+++ app/data/applist.json
@@ -1,1 +1,1 @@
-[{"appName": {"appName": "s", "filePath": "C:\\Users\\Shivendra Saurav\\Downloads\\s.jpg", "iconPath": "/public/images/s.jpg"}}]
+[]
